+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.6 (`2023-06-05`)
 - fix: use Retain reclaim policy for NFS storage
+- add: dnsChallenge via Let's Encrypt for Traefik
 
 # 0.0.5 (`2023-05-31`)
 - add: MetalLB as load balancer